[crashes] Move version transformer function to a module
